Changed: moved all styles interfaces in separated files
