+++ src/cli/lib/styles/index.ts
@@ -10,6 +10,7 @@ import { clipPath } from './clip-path';
 import { fill } from './fill';
 import { filters } from './filters';
 import { font } from './font';
+import { Parser } from './models';
 import { opacity } from './opacity';
 import { stroke } from './stroke';
 import { textAttributes } from './text-attributes';
@@ -23,10 +24,5 @@ const parsers: { [parser: string]: Parser } = {
   stroke,
   textAttributes,
 };
-
-export interface Parser {
-  name?: string;
-  parse: (...args) => string;
-}
 
 export default parsers;
