+++ src/cli/lib/styles/models/pattern.ts
@@ -6,13 +6,12 @@
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
 
-import { ClipPath, Parser } from './models';
-
-export const clipPath: Parser = {
-  name: 'clip-path',
-  parse: clipPathParser,
-};
-
-function clipPathParser(src: ClipPath): string {
-  return `url(#${src.ref})`;
+export interface Pattern {
+  meta: {
+    ux: {
+      uid: string;
+    };
+  };
+  width: number;
+  height: number;
 }
