+++ src/cli/lib/styles/models/text-attributes.ts
@@ -6,13 +6,7 @@
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
 
-import { ClipPath, Parser } from './models';
-
-export const clipPath: Parser = {
-  name: 'clip-path',
-  parse: clipPathParser,
-};
-
-function clipPathParser(src: ClipPath): string {
-  return `url(#${src.ref})`;
+export interface TextAttributes {
+  lineHeight: number;
+  paragraphAlign: string;
 }
