+++ src/cli/lib/styles/opacity.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
 
-import { Parser } from './index';
+import { Parser } from './models';
 
 export const opacity: Parser = {
   name: 'opacity',
