+++ src/cli/lib/styles/stroke.ts
@@ -6,17 +6,13 @@
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
 
-import { Fill, fillParser } from './fill';
-import { Parser } from './index';
+import { fillParser } from './fill';
+import { Parser, Stroke } from './models';
 
 export const stroke: Parser = {
   name: 'stroke',
   parse: strokeParser,
 };
-
-export interface Stroke extends Fill {
-  width: number;
-}
 
 function strokeParser(src: Stroke, parentElement, uuid, resources) {
   return fillParser(src, parentElement, uuid, resources) + `;stroke-width:${src.width}px`;
